+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 5.9-SNAPSHOT
 -------------
 
+ADDED:
+
+- new abstact class AbstractMutableComboBoxModel for a generic combobox model that can add and remove elements from the combobox dynamically
+
 Version 5.8
 -------------